requirements.txt
djangorestframework-simplejwt
djangorestframework
markdown
Django
django-filter